colab_backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
pydantic==2.6.1
aiofiles==23.2.1

# AI/ML Stack - PyTorch
torch==2.2.0
torchvision==0.17.0
diffusers==0.26.3
transformers==4.38.2
accelerate==0.27.2
controlnet-aux==0.0.10
xformers==0.0.24
opencv-python==4.9.0.80
numpy==1.26.4
safetensors==0.4.2
huggingface-hub==0.20.3

# MobileNet and Training
timm==1.0.20
scikit-learn==1.3.0
matplotlib==3.8.3
seaborn==0.13.2
tqdm==4.67.1

# Data Processing
pandas==2.1.4
datasets==2.16.1
requests==2.31.0

# Deployment
pyngrok==7.1.2

# Additional utilities
python-dotenv==1.0.1
psutil==7.1.0